Refactor Session and its ViewCommand code
